Add json minify to remove comments in JSON files
there is a common desire to put comments in a JSON files.
For example a license header at the top of the JSON file or
some other comments for clarfication of the contents.
Since comments is not supported in pure JSON, we need to remove
them before giving them to any kind of parser